current draft of cartogram implementation
something is not quite right with the effect of force in the cartogram
yet. It seems to be coming from the bottom left of the map, which is
incorrect.

+++ src/core/Cartogram.hpp
@@ -55,7 +55,7 @@ namespace GenGIS
 	private:
 		const int ysize;
 		const int xsize;
-		int valFudge;
+		double valFudge;
 		int areaFudge;
 		/** Copied directly from MapView. See class for more information. */
 		int m_dimension;
